feat(image-tools): F4.6.2 verbatim port image_analyze/edit/gen to dasclaw_image_tools
Verbatim port per ADR-129 §1.3 + ADR-156 §6.3 F4.6.2. Moves vision/image LLM
tools (image_analyze, image_edit, image_generate) from desktop-client builtin
into a new shared crate dasclaw_image_tools so future surfaces (TUI/CLI/MCP)
can consume the same implementation.

- New crate crates/dasclaw_image_tools (lib only, version 0.0.0-w6).
- git mv 3 files; pub fn media_type_from_path relocated to crate lib.rs.
- desktop-client builtin/mod.rs becomes a pub-use re-export, preserving
  existing consumer paths in tools/registry.rs.
- Cargo workspace + desktop-client Cargo.toml wired; blueprint sync
  planned-list entry removed.

No logic changes. 11 unit tests pass (path traversal / absolute-path
rejection / metadata / size & prompt validation).

Cross-cuts: ADR-114 类A (no new .ironclaw / IRONCLAW_BASE_DIR literals).
